edited bp_blog.create_post route to allow posts with no images

+++ app_package/bp_blog/routes.py
@@ -235,6 +235,8 @@ def create_post():
             dest = shutil.move(source, destination, copy_function = shutil.copytree) 
             logger_bp_blog.info(f"Destination path: {dest}") 
 
+            post_images_dir_name = ""
+
             # find root html file for post
             for file_name in os.listdir(dest):
                 
@@ -263,22 +265,26 @@ def create_post():
             print("*** END sanatize .fld ***")
             
 
+            if post_images_dir_name != "":
+                # beautiful soup to search and replace img src with {{ url_for('custom_static', ___, __ ,__)}}
+                # new_index_text = replace_img_src_jinja(os.path.join(new_blog_dir_fp,post_html_filename), unzipped_files_dir_name)
+                new_index_text = replace_img_src_jinja(os.path.join(new_blog_dir_fp,post_html_filename), post_images_dir_name)
+                if new_index_text == "Error opening index.html":# cannot imagine how this is possible, but we'll leave it.
+                    flash(f"Missing index.html? There was an problem trying to opening {os.path.join(new_blog_dir_fp,post_html_filename)}.", "warning")
+                    # return redirect(request.url)
+                    return redirect(url_for('bp_blog.blog_delete', post_id=new_blog_id))
 
-            # beautiful soup to search and replace img src with {{ url_for('custom_static', ___, __ ,__)}}
-            # new_index_text = replace_img_src_jinja(os.path.join(new_blog_dir_fp,post_html_filename), unzipped_files_dir_name)
-            new_index_text = replace_img_src_jinja(os.path.join(new_blog_dir_fp,post_html_filename), post_images_dir_name)
-            if new_index_text == "Error opening index.html":# cannot imagine how this is possible, but we'll leave it.
-                flash(f"Missing index.html? There was an problem trying to opening {os.path.join(new_blog_dir_fp,post_html_filename)}.", "warning")
-                # return redirect(request.url)
-                return redirect(url_for('bp_blog.blog_delete', post_id=new_blog_id))
+                # remove existing post_html_filename
+                os.remove(os.path.join(new_blog_dir_fp,post_html_filename))
 
-            # remove existing post_html_filename
-            os.remove(os.path.join(new_blog_dir_fp,post_html_filename))
+                # write a new index.html with new code that references images in image folder
+                index_html_writer = open(os.path.join(new_blog_dir_fp,post_html_filename), "w")
+                index_html_writer.write(new_index_text)
+                index_html_writer.close()
 
-            # write a new index.html with new code that references images in image folder
-            index_html_writer = open(os.path.join(new_blog_dir_fp,post_html_filename), "w")
-            index_html_writer.write(new_index_text)
-            index_html_writer.close()
+
+
+
 
             # delete compressed file
             shutil.rmtree(temp_zip_db_fp)
@@ -291,121 +297,6 @@ def create_post():
             logger_bp_blog.info(f"- filename is {new_post_dir_name} -")
 
 
-        #######################################################
-        # MARK: START
-        ######################################################
-        
-
-        # if request_files["zip_file_origin_word"].filename != "":
-        #     logger_bp_blog.info(f"- new_method -")
-
-        #     # get data from form
-        #     index_source = formDict.get('index_html_source')
-        #     if formDict.get('date_published'):
-        #         date_published_datetime = datetime.strptime(formDict.get('date_published'), "%Y-%m-%d")
-        #         print(f"date_published: {date_published_datetime}")
-        #     else:
-        #         date_published_datetime = datetime.utcnow()
-        #     post_zip = request_files["new_method"]
-        #     post_zip_filename = post_zip.filename
-
-        #     # create new_blogpost to get post_id number
-        #     new_blogpost = BlogPosts(user_id=current_user.id, date_published =date_published_datetime)
-        #     sess_users.add(new_blogpost)
-        #     sess_users.commit()
-
-        #     # create post_id string and 
-        #     new_blog_id = new_blogpost.id
-        #     new_post_dir_name = f"{new_blog_id:04d}_post"
-        #     new_blogpost.post_id_name_string = new_post_dir_name
-        #     sess_users.commit()
-
-        #     # save zip to temp_zip
-        #     temp_zip_db_fp = os.path.join(current_app.config.get('DB_ROOT'),'temp_zip')
-        #     if not os.path.exists(temp_zip_db_fp):
-        #         os.mkdir(temp_zip_db_fp)
-        #     else:
-        #         shutil.rmtree(temp_zip_db_fp)
-        #         os.mkdir(temp_zip_db_fp)
-            
-        #     post_zip.save(os.path.join(temp_zip_db_fp, secure_filename(post_zip_filename)))
-        #     zip_folder_name_nospaces = post_zip_filename.replace(" ", "_")
-
-
-        #     new_blog_dir_fp = os.path.join(current_app.config.get('DB_ROOT'), "posts", new_post_dir_name)
-        #     logger_bp_blog.info(f"- new_blog_dir_fp: {new_blog_dir_fp} -")
-
-        #     # check new_blog_dir_fp doesn't already exists -- This is a weird check but let's just leave it in....
-        #     if os.path.exists(new_blog_dir_fp):
-
-        #         # delete db entery
-
-        #         # delete db/posts/000_post dir
-
-        #         flash(f"This blog post is trying to build a directory to store post, but one already exists in: {new_blog_dir_fp}","warning")
-        #         return redirect(request.url)
-
-        #     # decompress uploaded file in temp_zip
-        #     with zipfile.ZipFile(os.path.join(temp_zip_db_fp, zip_folder_name_nospaces), 'r') as zip_ref:
-        #         print("- unzipping file --")
-        #         unzipped_files_foldername = zip_ref.namelist()[0]
-        #         unzipped_temp_dir = os.path.join(temp_zip_db_fp, new_post_dir_name)
-        #         print(f"- {unzipped_temp_dir} --")
-        #         zip_ref.extractall(unzipped_temp_dir)
-
-        #     logger_bp_blog.info(f"- decompressing and extracting to here: {os.path.join(temp_zip_db_fp)}")
-
-        #     unzipped_dir_list = [ f.path for f in os.scandir(unzipped_temp_dir) if f.is_dir() ]
-
-        #     # delete the __MACOSX dir
-        #     for path_str in unzipped_dir_list:
-        #         if path_str[-8:] == "__MACOSX":
-        #             shutil.rmtree(path_str)
-        #             print(f"- removed {path_str[-8:]} -")
-
-        #     # temp_zip path
-        #     source = unzipped_temp_dir
-        #     logger_bp_blog.info(f"- SOURCE: {source}")
-
-
-        #     # db/posts/0000_post
-        #     destination = os.path.join(current_app.config.get('DB_ROOT'), "posts")
-
-        #     dest = shutil.move(source, destination, copy_function = shutil.copytree) 
-        #     logger_bp_blog.info(f"Destination path: {dest}") 
-
-
-        #     # beautiful soup to search and replace img src with {{ url_for('custom_static', ___, __ ,__)}}
-        #     new_index_text = replace_img_src_jinja(os.path.join(new_blog_dir_fp,"index.html"))
-        #     if new_index_text == "Error opening index.html":
-        #         flash(f"Missing index.html? There was an problem trying to opening {os.path.join(new_blog_dir_fp,'index.html')}.", "warning")
-        #         # return redirect(request.url)
-        #         return redirect(url_for('bp_blog.blog_delete', post_id=new_blog_id))
-
-        #     # remove existing new_blog_dir_fp, index.html
-        #     os.remove(os.path.join(new_blog_dir_fp,"index.html"))
-
-        #     # write a new index.html with new_idnex_text
-        #     index_html_writer = open(os.path.join(new_blog_dir_fp,"index.html"), "w")
-        #     index_html_writer.write(new_index_text)
-        #     index_html_writer.close()
-
-
-        #     # delete compressed file
-        #     shutil.rmtree(temp_zip_db_fp)
-
-        #     # update new_blogpost.post_html_filename = post_id_post/index.html
-        #     new_blogpost.post_html_filename = os.path.join(new_post_dir_name,"index.html")
-        #     new_blogpost.title = get_title(os.path.join(new_blog_dir_fp,"index.html"), index_source)
-        #     sess_users.commit()
-
-        #     logger_bp_blog.info(f"- filename is {new_post_dir_name} -")
-
-
-
-        #######################################################
-        # MARK: END
-        ######################################################
 
         flash(f'Post added successfully!', 'success')
         # return redirect(url_for('bp_blog.blog_edit', post_id = new_blog_id))
